Branch out rediscloud as a separate shell plugin
