+++ plugins/rediscloud/api_key_test.go
@@ -1,4 +1,4 @@
-package redis
+package rediscloud
 
 import (
 	"testing"
